requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pandas>=2.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
httpx>=0.25.0
python-dateutil>=2.8.0
supabase>=2.0.0
reportlab>=4.0.0